Bonus déplacement des nains

+++ LibrarySmallWorld/GenerateurCarte.h
@@ -38,7 +38,7 @@ public:
 EXTERNC DLL GenerateurCarte* GenerateurCarte_New(int size, std::string PeupleA, std::string PeupleB);
 EXTERNC DLL void GenerateurCarte_Delete(GenerateurCarte* gc);
 
-EXTERNC DLL int* Cases_Destinations(std::string peuple, int rg, int * carte, int taille, int pm, int * posAdversaire, int nbAdversaires);
+EXTERNC DLL int* Cases_Destinations(std::string peuple, int rg, int * carte, int taille, double pm, int * posAdversaire, int nbAdversaires);
 
 class OperationSurCarte
 {
